pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pagr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "streamlit>=1.51.0",
    "python-dotenv>=1.2.1",
    "neo4j>=5.0.0",
    "pydantic>=2.5.0",
    "requests>=2.32.5",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
    "rich>=13.9.0",
    "pyvis>=0.3.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pagr"]
